rename lexer to tokenizer, provide lexer interface
Change name  of the token list  construction class to 'tokenizer'  as it is
more  fitting with  its  intended purpose.  Add base  lexer  class and  two
specific implementations of  the token iteration: traversing  full group of
tokens or walking  over sequence of token  IDs which is used as  a proxy to
the underlying token group.