test expected subset of tokens
Add `conf` field to the parse test specification -- it allows checking only
for  occurrence  of  specific  token  sets instead  of  requiring  a  large
copy-past (like I  did with the nested  token). It does not  provide a 100%
coverage for  particular output but  makes it  easier to test  for specific
lexer features such as indentation, nested sections etc.

- Added `maybe_field` function and several encode/decode overloads for YAML
  conversion. Now enumerations can be mapped to the fields a lot easier and
  proper exception message is provided. In the future it should be enhanced
  with 'did you mean' clarification based on the levenshtein edit distance.