Revert "transition to boost subgraph"
After  considering possible  features  (and existing  documentation of  the
subgraph library)  I've came to the  conclusion this is not  something that
would be useful  for the mind map  generation. Subtree is not  the only way
graph can be  split into clusters, encoding this specific  type of split in
the structure is not particularly useful.

Subgraph seems  to be a  fine library on  the surface, but  I had a  lot of
issues  making  the  compilation  work, documentation  is  lacking  on  the
details,  =stackoverflow=  examples are  a  bit  anemic  when it  comes  to
integrating  more complicated  features  ...  and so  on.  The usual  stuff
basically.

Instead I  will base mind  map generation on  the boost graph  entirely and
write custom generators based on this structure.

This reverts commit a3afa5f377e87dbb62da7bf6456241eb20c8ab70.