Don't use unicode when running tests in Qt
If the corpus test suite is  executed in the environment with IN_QT_RUN set
to 'true' formatting will automatically switch to the one that does not use
unicode symbols in  the printing or ANSI escape codes.  Instead it will use
ASCII characters  for all  the annotation  formatting (token  sequence diff
presentation). This makes it more readable  in the "test results" output of
the qt creator. Formatted output in QtCreator will look like this:

Lexed token mismatch
Given                                           Expected
~ [60]                             SameIndent "" <> "" SameIndent           [60]
- [61]                         ListItemStart "-" <
?                                                 > "--" ListItemStart      [61]
~ [62]                         SkipSpace "[SPC]" <> "[SPC]" SkipSpace       [62]