Introduce `value_domain<T>` struct
Instead of  relying on the  assortment of the freestanding  procedures that
have already caused some issues with  name resolution in the past I grouped
all  value  domain  introspection  logic   into  a  newly  introduced  type
`value_domain<T>`  that   is  modeled   closer  to  the   standard  library
`std::numeric_limit` and similar  structures. This also makes  it easier to
generate specializations to these enums based on the auto-generated code.

Right now  the code generated by  the python script is  not compatible with
the new  format and guile  script must  be updated accordingly,  to provide
high/low and other specifications.

Additionally,  enum  to/from string  conversion  was  also moved  into  the
structure  with  static fields  as  it  makes  it easier  to  automatically
generate these and group them together.