Create org spec object on demand
Instead of maintaining an org spec object as a static variable that is
linked to during compilation, use a separate function that creates a new
object as needed.

Right now this is causing spurious runtime issues in the corpus test
runnier and it is not clear what is the reason for this failure.