Enable text layouter back
Fix leftover debugging of the test suite failures. It is not clear why, but
the  duplicate text  fixture for  vector tests  caused memory  deallocation
issues in  the standard  library vector which  prevented the  text layouter
suite to be executed properly --  it double-freed memory or silently failed
to  allocated  it  in the  first  place  (`__p`  in  debugger was  zero  in
destructor).

Removed debugging  code from sem converter  -- writing of the  tree content
caused a 16 times slowdown in the execution for the notes file (24s vs 1.5s
for 40k-line file).