QtCreator dumper script for the sem ID values
DOD code  is a  lot less  accessible for  the debugger  visualizers because
"pointer-like" feature  of the ID (SemId  in this case) node  is present in
code, but not in the debugger. Luckily enough it is possible to implement a
custom debugging  helper for QtCreator  that will  format the value  in the
more human-readable manner. Now Sem ID values are printed like

eexport                @0x7ffffffeeea0     sem::SemIdT<sem::Export>
    [sem::SemId]       sem::SemId          sem::SemId
        ID             64424509441         unsigned long long
        Readable ID    "0_Export_0"        QString
        SEM             @0x55555c67f110    sem::Org
            loc         @0x55555c67f138    Opt<LineCol>
            original    @0x55555c67f128    OrgAdapter
            parent      sem::SemId         sem::SemId
            subnodes    @0x55555c67f148    Vec<sem::SemId>