org: add sem placement context information
Additional contextual information about node placement with respect to
the parent tree. In some cases it contains data that cannot be directly
restored after the sem convert (for example, determining whether
paragraph is placed in the title or in the body is not possible simply
via 'parent' relation, it must be done by traversing the subtrees of a
parent node and determining where exactly the node is located -- pretty
much replicating the sem converter logic) so it is easier to record it
in a field.