NLP: Don't send all HTTP req. at once, wait to process
Sending all requests at once overwhelmed the Langtool server and caused
it to cancel operations. Limiting max number of requests in flight to 20
solved the issue. With caching and max number limit it is possible to
process 6.5k line document in 70 seconds (initial pass) and then 4.3
seconds (subsequent passes with caching)