codegen: remove guile definitions
In the end guile provide to  be a very questionable solution for describing
data  in  the manner  I  used  here. It  does  not  provide a  higher-level
bindings, has  an abysmal level  of reporting  for the backtraces,  hard to
understand after a short period of time  (originally I picked it due to the
'data-as-code' appeal, but it is neither needed here, nor desired)

Before choosing python  I tried to use  racket, but it ended  up in failure
because  of  completely  inaccessible   documentation  and  barely  legible
instructions  on  how  to  embed  it  into  a  C  program.  Compared  to  a
`boost::python` it is a complete waste of time.

Another  language that  could have  been used  for this  is lua,  with data
tables and  fast loading  times, but I'm  neither running  into performance
bottlenecks (the code /is/ slow, taking about a second to generate mere six
files,   but  it   is  still   thousands  of   times  faster   than  manual
implementation).