codegen: remove c++ implementation
Dropped  boost  python helpers  together  with  the code  generator  logic.
Converters  might  be  used  later  on  to  interop  with  embedded  python
interpreter, but at the moment this is not a priority.

First note  on transition  : boost  python is  an acceptable  solution, but
severe  lack  of  examples  in   documentation  when  it  comes  to  making
non-trivial things  (apparently passing  ~std::vector<T>~ is  a non-trivial
thing that  might need  to be  researched additionally),  it does  not have
wrappers for  interop on standard  containers. =pybind11= is almost  a full
drop-in replacement but it has not  caused a single memory-related issue as
of yet.

Second note  on transition:  depends on  the use  case, but  embedding less
common languages such as guile or  lua is almost always guaranteed to cause
issues with  insufficient libraries, missing  documentation or the  need to
hack boilerplate elements on top of each other. In this regard python is
the best solution in terms of available features.