Generate `[]` operator for exported IDs
This is an  intermediate solution for getting subnode at  point for the org
IDs -- later on it would be  better to generate proper cxx record hierarchy
and export it directly, mirroring how  main types are defined. There should
be a base `PyOrgId` that is  exported as `SemNode` or something similar and
other type are annotated as derived.

This would allow to flow down other methods in hierarchy, like attached and
statements.

At the moment `sem::SemId` return type for index is not wrapped, it will be
added next.