rename non-mutating 'visit' to 'eval'
The naming of different CRTP methods for the visitor was confusing --
there were 'visit' functions that accepted results and were allowed to
mutate them and ones that were not allowed to do this. Now it is 'eval'
for things that return value and 'visit' for things that mutate passed
result argument.