Initial outline of the py-tex exporter
There are  lots of questionable  places in  the implementation and  I don't
particularly  like   how  many  different   hooks  are  exposed   from  the
exporter (maybe this is just a matter of formatting them though), but
overall the approach seems to be working.

There are several  pitfalls in the python  visitor UX writing --  (1) it is
not  clear  what methods  can  be  overridden, there  is  no  full list  or
auto-complete for this case (they can  be annotated as virtual or something
like that?), (2) when  a method is not called as expected  you need to look
into a  convoluted trace, (3)  because of the class  decorator autocomplete
for some fields is not working properly.

Otherwise, it seems  like this approach is going to  work perfectly. Export
seed  might be  an  issue, but  for  things  like HTML  and  Latex I  think
incremental conversion  on file level will  be sufficient to address  it in
most cases. And at the moment I don't even know timings anyway.