WIP Generate ~static_cast~ wraps for pybind11
All  exposed functions  should be  wrapped  in ~static_cast~  to allow  for
easier disambiguation when overload comes in.  It was not a problem because
for some  reason sub-variant getter  functions got compiled  properly (most
likely because the argument is non-const there).

Also fixed several linter errors in  the main codegen file and moved =yapf=
to the file root.