Dump emacs lisp org-mode parsing as pydantic model
I  have never  considered  writing  an emacs  wrapper  solution instead  of
writing my own library and this  short experiment only confirmed my initial
thoughts. Mutually  recursive data,  Emacs-specific S-expressions  that are
not easily mappable  to JSON, property lists with  unclear structural rules
and so on and so forth. Hard to interface with, hard to debug.