Use address sanitizer
There is no better  time to enable this option than while  I'm on the topic
of rewriting everything  from scratch. Previously, it was too  hard to turn
it on because  of the unexpected failures  in the hashtag nodes,  but now I
can track down the new edits and debug them as they appear.