Looks like ~ExternalProject~ is not for deps
Adding something with external project does not create a target that can be
integrated with the regular build process and I'm really puzzled as to what
this thing is  needed for in this  case. In any scenario, it  is not needed
for *my*  case of dependency management,  so it can drop  down to /dev/null
for all I care. But the project  builds now, that is the important part. No
linker errors from the standard library, clang also worked fine etc.