Initial config for sphinx doc generator
WIP configuration of the sphinx documentation generator. I have no idea why
this  POS  requires  so  much  boilerplate to  do  something  as  basic  as
generating documentation from  all ~.py~ files in  the directory (=doxygen=
can do this with only some  superficial configuration that is done from GUI
and that's it). In any case, I  added the builder task to compile the whole
list of modules.