repo: less mutable state in commit iteration
Refactoring  to  reduce  amount  of  mutable state  stored  in  the  commit
iteration --  identical implementations ~apply()~  approach was not  a good
idea, so with separate return types the  state can be stored in the current
scope of iteration instead of explicit  set of fields. Make the compiler do
all the cleanup.