repo: simplify file rename track adjustment
Current active  file track  bookkeeping had  an unnecessary  component that
tracked  known file  renames. This  made the  change iteration  state model
misalignment with the  main repository -- in the real  repository file move
did not  leave any extra residual  information, but in the  cxx code replay
had this  a->b->c link system.  Because of this  link it was  impossible to
handle the following repo use case:

1. Create file A
2. Move A -> B
3. Create A again
4. Edit B

Editing 'B'  and then checking for  the file modification state  brought up
the content of 'A' mixed up, because  there was a reference from 'B' to 'A'
which now collided.

Moving  file tracks  in the  tables for  real is  both closer  to the  real
scenario (which is important here) and makes implementation simpler.

Now all repo-based tests are passing.