org: wip indented list lexing
The new code seems like an overengineered garbage, but the previous one was
pretty much in  the same ballpark. Now at least  the operation is happening
on a much larger blocks of content and it should be easier to support cases
like

```
Some content

  - First list item with leading indentation

    - Indent things
    - #+begin_src
Whatever, this is legal org-mode
- Hello list-ish
      #+end_src
      - Sub-item

Some more content
```

At least I hope it would  be easier because of multiple data transformation
steps for lex recombination. /At some stage/ in the transform the IR should
be suitable for the task.