org: use __builtin_X for report locations
Reduce  usage of  macros in  the tracer  for the  parser and  tokenizer and
instead rely on the clang __builtin_line and __builtin_function elements to
provide the  location. ~std::source_location~  is not  fully stable  in the
current  implementation, but  later  on  it would  be  possible  to use  it
instead.