org: protobuf codegen implementation checkpoint
Protobuf grammar is  constructed properly, now working out  issues with the
cxx code  generation. Fully automatic  conversion of the cxx  structures to
protobuf data does not seem to  be doable with reflection because of highly
irregular structure  of the generated  API --  some things are  called with
~set_~ (integers, enums), some are with ~mutable_~ (sequences). Enum fields
do not have ~mutable_~ accessor.