org: protobuf grammar API for serde and fuzzing
Use protobuf to  describe serialization and deserialization  format for org
documents in binary  form for caching, interchange,  formally described API
in  a machine-readable  form and  most importantly  compatible with  google
fuzztest implementation  for automatic random document  generation, which I
plan to use for testing later.

Next step is to implement  protobuf de-serialization form, should be easier
with most of the writer components intact. Then, exporter from the org-mode
document to the text form and finally bring all these three pieces together
in an automatic  and semi-automatic fuzz testing  solution to automatically
find bugs and edge cases in the parser.