org: make block arguments optional, fixes serde
With  this commit  protobuf in-out  serialization test  is passing  and the
constructed  sem tree  for  the document  match the  original  one down  to
ordering and values of each node. IDs are not fully stable here and this is
something that need to be investigated as well, but I believe this is cased
by some  content that is  temporarily dropped from  the tree at  some point
during parsing.