org: drop the state tracking from the tokenizer
Moving more context-dependent logic to the tokenizer. There is still a need
for  more tricky  parts of  the token  conversion logic,  partially because
there is no  look-back support on the re/flex lexer,  but otherwise I think
the  time  has  come  to  finally kill  this  overengineered  garbage  that
permeated  the token  processing and  start implementing  the logic  in the
parser, where it actually makes sense to do context-aware decisions.

This commit is more of a headshot-style refactoring, in the next steps if I
will try  to look what intensities  can still be  put to some use  and what
parts will  go straight into  the graveyard  of other stillborn  ideas that
somehow lived through all the rewrites.