lex: sub-token lexer rules
Org-mode tokens often need to be  further separated into sub-tokens, but in
order to do this with a single  lexer lots of complex state transitions are
necessary,  to keep  track of  the current  state. With  sub-lexer support,
lexing of things  like ~#+tblfm:~ can be done much  simpler. Take the whole
line in and then subdivide it as needed.