repo: drop fuzztest for testing
After working through a set of real  test files I do not see fuzztest could
have  caught  any  meaningful  portion  of these  bugs  and  it  lacks  the
flexibility in order  to engineer the test strategy  generator that /would/
be able  to do so.  Later on I  will use the  python driver to  create test
documents for checking.