editor: use lexbor and HTML exporter
=QTextDocument= does not  provide enough tools to cleanly  map the incoming
rich text to  the org node structure, exporter to  the =QTextDocument= also
does not  provide enough control over  the resulting text structure  as the
whole thing is written as a stream of text fragments.

Directly  generating  HTML results  provides  a  simpler logic  of  mapping
between org-mode and what is displayed in the QML side.