org: add missing perfetto static storage defs
`<sem/perfetto_org.hpp>`                                                and
`<hstd/wrappers/perfetto_aux_impl_template.hpp>`   need   to  be   included
exactly once  in every compiled  binary -- either in  the main file,  or in
some  cpp configuration,  where they  define static  storage for  the trace
events. For  unknown reason it compiled  in the debug build  and in earlier
releases, but building it now causes the linker errors to finally surface.