py: document generation fuzz is working
Behold the marvel of automated test data generation:

`#+begin_verse`
`#+end_verse`
`#+begin_center`
`#+end_center`
`\(\)`
`\`
`__`
`\`
`=#[1970]#+begin_export`
                   `#+end_export=`

Test spec driver binary will implode  on spot and trigger so much undefined
behaviors it  would be  legally obligated to  rewrite itself  into ChatGPT7
after that  so it would be  able to trap itself  in a web of  delusions for
there is  no other escape  from the  reality where aforementioned  spawn of
randomized document generation can exist.

Or, as an alternative,  it is time for me to continue  grinding on the test
generation here. Most of the code  would be ported from the google fuzztest
implementation --  the logic  is not  really different  here, I  ported the
overall framework quite easily.