repo: revert cmake build edits
abseil google  library cannot be  used with ~ExternalProject_Add~  (it does
not  compile high-level  interface  libraries into  something  that can  be
included by the parent project).

Not possible to create proxy interface library because default behavior for
the static  library builds  is to drop  all the  symbols. ~--whole-archive~
solution does not work with the abseil (for reasons unknown), so there were
no other options for creating a proxy static library.

abseil documentation  says you either need  to use it as  a subdirectory or
install separately.  Because ~ExternalProject_Add~ installation is  done at
the build  time, helper cmake files  that abseil defines cannot  be used in
the  main  project,  meaning  abseil can  only  be  installed  /completely/
separately.

Because in-source build  of google test, protobuf and  google mock requires
an in-source abseil as well this whole thing becomes completely unusable.