cli: test JS visualization
I absolutely  detest doing anything  with JS and especially  its ecosystem,
but it  looks like  there is  almost no  alternative to  d3.js in  terms of
supported visualization features and simplicity  of interactions. So the JS
code for visualization is  here to stay, which means it  must be tested and
verified.

Just look at this garbage commit though,  828 lines of lock file, 150 lines
of gitignore to counter-hack all this disgusting vomit that node spewed all
over the code, and who knows what else is coming here.