doc: generate documentations for tests and scripts
Add test  and script directories into  the list of files  processed for the
documentation. Changed  field declaration in  the program state  handler --
tree-sitter generates very  strange AST for this one, and  the macro itself
is probably too convoluted to keep it  in the code anyway. It saves several
field copy-pastes, but creates this weird construct in the code instead.