doc: remove coverage collector V1
CXX code coverage collection implementation based on the ~llvm-cov~ exports
to JSON  is too slow  and convoluted to use  on a regular  basis. Exporting
JSON from  the current test  coverage creates  30GB of temporary  files and
takes 20  minutes to  export all  the coverage information.  This is  not a
workable  solution.  V2 solution  should  use  LLVM  API for  working  with
coverage  files directly  and  export  a unified  sqlite  database for  all
coverage runs registered for a particular binary file.