editor: remove POC for GUI drag and drop
Node movement is better implemented  using shortcuts and emacs-like command
panel. In the end there would be a lot more different operations for moving
trees -- refiling  to different elements, promoting  and demoting elements,
so it is easier to just go with the command approach from the start instead
of working on the GUI drag and drop.