editor: generate test class binaries automatically
QTest has a very unfortunate limitation  -- every single test class must be
executed from a separate binary. CLI  does not support multiple classes, qt
creator does not support multiple  classes either. Most common solution you
can find on the  internet is to ~OR~ together many  results of the ~qExec~,
but that (1) does  not show multiple classes on the CLI,  (2) does not show
multiple classes on the qtcreator.

I added a JSON file with description of test class name and the header file
where test is defined. CMake reads this file and automatically creates many
separate boilerplate binaries that can be easily executed one by one.

I  used  JSON to  later  read  it into  the  qt  configuration. It  is  not
implemented now, but I will handle this later on.