mmap: first test of the adaptagrams library
Slowly  making sense  of the  constraint solution  and on  how it  might be
applied for the  mind map rendering later on. Also  re-enabled the graphviz
wrappers that  I commented out  due to the  libc++ library issues.  Since I
switched to  the system-provided standard  library back again this  part of
the project can be enabled without any issues.