mmap: store mutual pointers in the store/tree
I wrote the document  tree to mostly use it as a  structural outline of the
document with empty "boxes" where I  put specific content of the paragraph,
subtree, list  item etc.  So the  collection of trees,  once read  from the
file/string, is  not going  to change  too much  unless there  are explicit
moves around the structure. But in  this case the whole tree structure will
move accordingly.

To speed  up look up from  boxes to the  containing trees I added  a lookup
table. Since the trees are already  tied to the specific store instance via
IDs I also added pointer to the originating store.