mmap: access data fields directly for graphviz
name-based property  retrieval is /not/  the universal way for  getting the
data in graphviz, it is only  suitable for user-defined properties. All the
values that graphviz expects to work with are stored in the graph/node/edge
info object and could be accessed using  a hacky set of macros like ~GD_bb~
for bounding box.