mmap: deduplicate node document construction
It is necessary to create text document  for each graph node in two places:
during layout  phase and  then during  rendering itself.  The first  one is
needed purely for  the node size, the  second one is for the  text. In both
cases the documents must match exactly,  otherwise it would create a layout
issues.

The solution  with callback creates a  data flow that goes  other way (from
scene view to the proxy model), but  because the rendering part will have a
final  say  in  the  node  size (different  rendering  styles  can  require
different margins etc) this seems reasonable.