mmap: recount rows on the source model change
For reasons that I have no hope of ever understanding, calling ~rowCount()~
when the  layout changes makes  the graph filter actually  synchronize with
the source model.

The test  ~testGraphLayoutUpdateSignals~, when executed with  proxy filter,
failed because  the source model  reported 6 rows in  the ~getFullLayout()~
and a result it tried to get an invalid index.