editor: put tree modification logic to OrgTreeNode
Other parts of the editor can also request structural modifications for the
org tree node  -- inserting subtree, moving elements around  and so on. Org
store is the central element that manages ownership of all the roots loaded
in the  editor, not  the document  model. So  it makes  sense to  move tree
editing operations  in there, and then  broadcast all the changes  for each
tree from the  store. Each separate document model will  handle edits under
it's root.