gui: move persistent zoom state to class
`update()` method  creates a lot  of intermediate variables, most  of which
are not used  after the method has  finished, but there are a  few that are
captured in  the anonymous functions and  outlive the scope of  the method.
Moving  all of  them to  separate 'state'  class to  make this  a bit  more
explicit.