org: ignore node location for immutable AST hash
Node location is not really a part of the "content" and each individual
node is likely to have a unique location within a single document, but this
does not hold true across multiple documents.

All of this makes the location field an unnecessary complication when it
comes to node interning and comparison, so it is better to drop it
altogether.