repo: test cpack build with minimum deps use
The only two libraries that are almost guaranteed to be missing from the
package manager or the system installation is the reflex lexer and the lexy
parser libraries. Boost mp11 and describe are likely to be present as
headers, although for antiquated systems like `ubuntu:latest` the versions
might be lagging a century or two behind, and I need a much more recent
1.85 version of the boost headers for these libraries. So these two might
also be included.