gui: add initial API for diagram editing
The editing  operation can be implemented  as a group of  "edit" operations
sent  from the  user-visible  part of  the app.  The  diagram context  will
process the  edits, and create a  new diagram tree, then  send signals with
the  changes --  using  the same  mechanism as  the  undo/redo for  history
switches.